feat(app): route home through /start with dismissible start page
The fresh-install start page lived at `/`, and its "display this start
page next time" checkbox had no effect — nothing consumed the cookie it
set, so unchecking it did nothing.

Split the concerns: `/start` now owns the landing content and always
renders it, while `/` becomes a pure router. Unless the visitor has
dismissed the start page, `/` redirects to `/start`; otherwise it sends
authenticated users to `/backend` and everyone else to `/login`.

The checkbox persists a one-way `start_page_dismissed` cookie (replacing
the inert `show_start_page` flag): unchecking dismisses the page and
leaves via `/`, and re-checking it on `/start` clears the dismissal.

Mirrored into the create-app template so scaffolded apps get the same
behavior.
